+++ src/jest-util/src/init.lua
@@ -7,6 +7,9 @@
  * LICENSE file in the root directory of this source tree.
  ]]
 
+local LuauPolyfill = require(script.Parent.LuauPolyfill)
+local Object = LuauPolyfill.Object
+
 local exports = {}
 
 exports.clearLine = require(script.clearLine).default
@@ -14,24 +17,37 @@ exports.createDirectory = require(script.createDirectory).default
 local ErrorWithStackModule = require(script.ErrorWithStack)
 exports.ErrorWithStack = ErrorWithStackModule.default
 export type ErrorWithStack = ErrorWithStackModule.ErrorWithStack
-exports.installCommonGlobals = require(script.installCommonGlobals).default
-exports.interopRequireDefault = require(script.interopRequireDefault).default
+-- ROBLOX deviation: need to execute the module explicitly
+exports.installCommonGlobals = require(script.installCommonGlobals)().default
+-- ROBLOX deviation not ported as it doesn't seem necessary in Lua
+-- exports.interopRequireDefault = require(script.interopRequireDefault).default
 exports.isInteractive = require(script.isInteractive).default
 exports.isPromise = require(script.isPromise).default
 exports.setGlobal = require(script.setGlobal).default
 exports.deepCyclicCopy = require(script.deepCyclicCopy).default
--- ROBLOX deviation: not ported as it doesn't seems necessary in Lua
--- exports.convertDescriptorToString = require(script.convertDescriptorToString).default
-exports.specialChars = require(script.specialChars)
--- ROBLOX deviation: not ported as it doesn't seems necessary in Lua
+exports.convertDescriptorToString = require(script.convertDescriptorToString).default
+local specialCharsModule = require(script.specialChars)
+Object.assign(exports, specialCharsModule)
+-- ROBLOX deviation START: additional assignments for Lua type inferrence to work
+exports.ARROW = specialCharsModule.ARROW
+exports.ICONS = specialCharsModule.ICONS
+exports.CLEAR = specialCharsModule.CLEAR
+-- ROBLOX deviation END
+-- ROBLOX deviation: not ported as it doesn't seem necessary in Lua
 -- exports.replacePathSepForGlob = require(script.replacePathSepForGlob).default
--- ROBLOX deviation: not ported as it doesn't seems necessary in Lua
+-- ROBLOX deviation: not ported as it doesn't seem necessary in Lua
 -- exports.testPathPatternToRegExp = require(script.testPathPatternToRegExp).default
 exports.globsToMatcher = require(script.globsToMatcher).default
-exports.preRunMessage = require(script.preRunMessage)
+local preRunMessageModule = require(script.preRunMessage)
+Object.assign(exports, preRunMessageModule)
+-- ROBLOX deviation START: additional assignments for Lua type inferrence to work
+exports.print = preRunMessageModule.print
+exports.remove = preRunMessageModule.remove
+-- ROBLOX deviation END
 exports.pluralize = require(script.pluralize).default
 exports.formatTime = require(script.formatTime).default
--- ROBLOX deviation: not ported as it doesn't seems necessary in Lua
+-- ROBLOX deviation START: not ported as it doesn't seem necessary in Lua
 -- exports.tryRealpath = require(script.tryRealpath).default
-exports.requireOrImportModule = require(script.requireOrImportModule).default
+-- exports.requireOrImportModule = require(script.requireOrImportModule).default
+-- ROBLOX deviation END
 return exports